+++ src/SPLwithGauss.java
@@ -1,5 +1,5 @@
 public class SPLwithGauss {
-    public static boolean isRowManySolution(float[] row) {
+    public static boolean isRowManySolution(double[] row) {
         int col = 0;
         boolean isManySolution = true;
 
@@ -13,8 +13,8 @@ public class SPLwithGauss {
         return isManySolution;
     }
 
-    public static void SPLGauss(float[][] matrix) {
-        float[][] matrixHasil;
+    public static void SPLGauss(double[][] matrix) {
+        double[][] matrixHasil;
         int solusi = 1, idx, row, col;
         boolean isAllNol = true, isSquare = true;
 
@@ -38,7 +38,7 @@ public class SPLwithGauss {
         }
 
         if (solusi == 1) {
-            float solusiX[] = new float[row-1];
+            double solusiX[] = new double[row-1];
             int idxUnik, idxHasilUnik;
 
             for (idxUnik = row - 1; idxUnik >= 0; idxUnik--) {
@@ -57,7 +57,7 @@ public class SPLwithGauss {
         } else if (solusi == 2) {
             System.out.println("Tidak ada Solusi");
         }  else if (solusi == 3) {
-            float[] solusiX = new float[col];
+            double[] solusiX = new double[col];
             int idxRow, idxCol;
             
             for (idxRow = 0; idxRow < row; idxRow++) {
